Introduce unit tests for peer base list from P2P peer directory
