+++ elements/lisk-p2p/test/unit/peer_directory/new_list.ts
@@ -48,174 +48,6 @@ describe('newPeer', () => {
 		});
 	});
 
-	describe('#addPeer', () => {
-		let newPeersObj: NewList;
-		const samplePeers = initializePeerInfoList();
-
-		beforeEach(async () => {
-			newPeersObj = new NewList(newPeerConfig);
-			newPeersObj.addPeer(samplePeers[0]);
-		});
-
-		it('should add the incoming peer when it does not exist already', async () => {
-			expect(newPeersObj.getPeer(samplePeers[0])).eql(samplePeers[0]);
-		});
-
-		it('should not add the incoming peer if it exists', async () => {
-			expect(newPeersObj.addPeer(samplePeers[0]))
-				.to.be.an('object')
-				.haveOwnProperty('isAdded').to.be.false;
-		});
-	});
-
-	describe('#getNewPeersList', () => {
-		const samplePeers = initializePeerInfoList();
-		let newPeersObj: NewList;
-		let newPeersArray: ReadonlyArray<P2PPeerInfo>;
-
-		beforeEach(async () => {
-			newPeersObj = new NewList(newPeerConfig);
-			newPeersObj.addPeer(samplePeers[0]);
-			newPeersObj.addPeer(samplePeers[1]);
-			newPeersObj.addPeer(samplePeers[2]);
-			newPeersArray = newPeersObj.peersList();
-		});
-
-		it('should return new peers list', async () => {
-			const expectedNewPeersArray = [
-				samplePeers[0],
-				samplePeers[1],
-				samplePeers[2],
-			];
-			expect(newPeersArray).to.have.members(expectedNewPeersArray);
-		});
-	});
-
-	describe('#removePeer', () => {
-		let newPeersObj: NewList;
-		const samplePeers = initializePeerInfoList();
-
-		beforeEach(async () => {
-			newPeersObj = new NewList(newPeerConfig);
-			newPeersObj.addPeer(samplePeers[0]);
-			newPeersObj.addPeer(samplePeers[1]);
-		});
-
-		it('should remove the peer from the incoming peerInfo', async () => {
-			newPeersObj.removePeer(samplePeers[0]);
-			expect(newPeersObj.getPeer(samplePeers[0])).to.be.undefined;
-		});
-	});
-
-	describe('#getPeer', () => {
-		let newPeersObj: NewList;
-		const samplePeers = initializePeerInfoList();
-
-		beforeEach(async () => {
-			newPeersObj = new NewList(newPeerConfig);
-			newPeersObj.addPeer(samplePeers[0]);
-			newPeersObj.addPeer(samplePeers[1]);
-		});
-
-		describe('when peer exists in the triedPeers peerMap', () => {
-			it('should get the peer from the incoming peerId', async () => {
-				expect(newPeersObj.getPeer(samplePeers[0]))
-					.to.be.an('object')
-					.and.eql(samplePeers[0]);
-			});
-		});
-
-		describe('when peer does not exist in the triedPeers peerMap', () => {
-			const randomPeer = initializePeerInfoList()[2];
-			it('should return undefined for the given peer that does not exist in peerMap', async () => {
-				expect(newPeersObj.getPeer(randomPeer)).to.be.undefined;
-			});
-		});
-	});
-
-	describe('#updatePeer', () => {
-		let newPeersObj: NewList;
-		const samplePeers = initializePeerInfoList();
-
-		beforeEach(async () => {
-			newPeersObj = new NewList(newPeerConfig);
-			newPeersObj.addPeer(samplePeers[0]);
-			newPeersObj.addPeer(samplePeers[1]);
-		});
-
-		describe('when trying to update a peer that exist', () => {
-			it('should update the peer from the incoming peerInfo', async () => {
-				let updatedPeer = {
-					...samplePeers[0],
-					height: 0,
-					version: '1.2.3',
-				};
-
-				const success = newPeersObj.updatePeer(updatedPeer);
-				expect(success).to.be.true;
-				expect(newPeersObj.getPeer(samplePeers[0])).to.be.eql(updatedPeer);
-			});
-		});
-
-		describe('when trying to update a peer that does not exist', () => {
-			it('should return false when the peer does not exist', async () => {
-				let updatedPeer = {
-					...samplePeers[2],
-					height: 0,
-					version: '1.2.3',
-				};
-
-				const success = newPeersObj.updatePeer(updatedPeer);
-				expect(success).to.be.false;
-			});
-		});
-	});
-
-	describe('#findPeer', () => {
-		let newPeersObj: NewList;
-		const samplePeers = initializePeerInfoList();
-
-		beforeEach(async () => {
-			newPeersObj = new NewList(newPeerConfig);
-			newPeersObj.addPeer(samplePeers[0]);
-			newPeersObj.addPeer(samplePeers[1]);
-		});
-
-		describe('when the peer exist', () => {
-			it('should find the peer from the incoming peerInfo', async () => {
-				const peer = newPeersObj.getPeer(samplePeers[0]);
-				expect(peer).eql(samplePeers[0]);
-			});
-		});
-
-		describe('when the peer does not exist', () => {
-			it('should return false when the peer does not exist', async () => {
-				const success = newPeersObj.updatePeer(samplePeers[2]);
-				expect(success).to.be.false;
-			});
-		});
-	});
-
-	describe('#failedConnectionAction', () => {
-		let newPeersObj: NewList;
-		const samplePeers = initializePeerInfoList();
-
-		beforeEach(async () => {
-			newPeersObj = new NewList(newPeerConfig);
-			newPeersObj.addPeer(samplePeers[0]);
-			newPeersObj.addPeer(samplePeers[1]);
-		});
-
-		describe('when the peer exist and applied failedConnectionAction', () => {
-			it('should delete the peer and for the second call it should return false', async () => {
-				const success1 = newPeersObj.failedConnectionAction(samplePeers[0]);
-				expect(success1).to.be.true;
-				const success2 = newPeersObj.failedConnectionAction(samplePeers[0]);
-				expect(success2).to.be.false;
-			});
-		});
-	});
-
 	describe('#evictionBasedOnRandomSelection', () => {
 		const newPeerConfig = {
 			peerBucketSize: 2,
