+++ elements/lisk-p2p/test/unit/peer_directory/tried_list.ts
@@ -41,155 +41,6 @@ describe('triedPeer', () => {
 		});
 	});
 
-	describe('#addPeer', () => {
-		let triedPeersObj: TriedList;
-		const samplePeers = initializePeerInfoList();
-
-		beforeEach(async () => {
-			triedPeersObj = new TriedList(triedPeerConfig);
-			triedPeersObj.addPeer(samplePeers[0]);
-		});
-
-		it('should add the incoming peer if it does not exist already', async () => {
-			expect(triedPeersObj.getPeer(samplePeers[0])).eql(samplePeers[0]);
-		});
-
-		it('should not add the incoming peer if it exists', async () => {
-			expect(triedPeersObj.addPeer(samplePeers[0]))
-				.to.be.an('object')
-				.haveOwnProperty('isAdded').to.be.false;
-		});
-	});
-
-	describe('#gettriedPeersObj', () => {
-		const samplePeers = initializePeerInfoList();
-		let triedPeersObj: TriedList;
-		let triedPeersArray: ReadonlyArray<P2PDiscoveredPeerInfo>;
-
-		before(async () => {
-			triedPeersObj = new TriedList(triedPeerConfig);
-			triedPeersObj.addPeer(samplePeers[0]);
-			triedPeersObj.addPeer(samplePeers[1]);
-			triedPeersObj.addPeer(samplePeers[2]);
-			triedPeersArray = triedPeersObj.peersList() as ReadonlyArray<
-				P2PDiscoveredPeerInfo
-			>;
-		});
-
-		it('should return tried peers list', async () => {
-			const expectedTriedPeersArray = [
-				samplePeers[0],
-				samplePeers[1],
-				samplePeers[2],
-			];
-			expect(triedPeersArray).to.have.members(expectedTriedPeersArray);
-		});
-	});
-
-	describe('#removePeer', () => {
-		let triedPeersObj: TriedList;
-		const samplePeers = initializePeerInfoList();
-
-		beforeEach(async () => {
-			triedPeersObj = new TriedList(triedPeerConfig);
-			triedPeersObj.addPeer(samplePeers[0]);
-			triedPeersObj.addPeer(samplePeers[1]);
-		});
-
-		it('should remove the peer from the incoming peerInfo', async () => {
-			triedPeersObj.removePeer(samplePeers[0]);
-			expect(triedPeersObj.getPeer(samplePeers[0])).to.be.undefined;
-		});
-	});
-
-	describe('#getPeer', () => {
-		let triedPeersObj: TriedList;
-		const samplePeers = initializePeerInfoList();
-
-		beforeEach(async () => {
-			triedPeersObj = new TriedList(triedPeerConfig);
-			triedPeersObj.addPeer(samplePeers[0]);
-			triedPeersObj.addPeer(samplePeers[1]);
-		});
-
-		describe('when peer exists in the triedPeers peerMap', () => {
-			it('should get the peer from the incoming peerId', async () => {
-				expect(triedPeersObj.getPeer(samplePeers[0]))
-					.to.be.an('object')
-					.and.eql(samplePeers[0]);
-			});
-		});
-
-		describe('when peer does not exist in the triedPeers peerMap', () => {
-			const randomPeer = initializePeerInfoList()[2];
-			it('should return undefined for the given peer that does not exist in peerMap', async () => {
-				expect(triedPeersObj.getPeer(randomPeer)).to.be.undefined;
-			});
-		});
-	});
-
-	describe('#updatePeer', () => {
-		let triedPeersObj: TriedList;
-		const samplePeers = initializePeerInfoList();
-
-		beforeEach(async () => {
-			triedPeersObj = new TriedList(triedPeerConfig);
-			triedPeersObj.addPeer(samplePeers[0]);
-			triedPeersObj.addPeer(samplePeers[1]);
-		});
-
-		describe('when trying to update a peer that exist', () => {
-			it('should update the peer from the incoming peerInfo', async () => {
-				let updatedPeer = {
-					...samplePeers[0],
-					height: 0,
-					version: '1.2.3',
-				};
-
-				const success = triedPeersObj.updatePeer(updatedPeer);
-				expect(success).to.be.true;
-				expect(triedPeersObj.getPeer(samplePeers[0])).to.be.eql(updatedPeer);
-			});
-		});
-
-		describe('when trying to update a peer that does not exist', () => {
-			it('should return false when the peer does not exist', async () => {
-				let updatedPeer = {
-					...samplePeers[2],
-					height: 0,
-					version: '1.2.3',
-				};
-
-				const success = triedPeersObj.updatePeer(updatedPeer);
-				expect(success).to.be.false;
-			});
-		});
-	});
-
-	describe('#findPeer', () => {
-		let triedPeersObj: TriedList;
-		const samplePeers = initializePeerInfoList();
-
-		beforeEach(async () => {
-			triedPeersObj = new TriedList(triedPeerConfig);
-			triedPeersObj.addPeer(samplePeers[0]);
-			triedPeersObj.addPeer(samplePeers[1]);
-		});
-		describe('when the peer exist', () => {
-			it('should find the peer from the incoming peerInfo', async () => {
-				const peer = triedPeersObj.getPeer(samplePeers[0]);
-				expect(peer).eql(samplePeers[0]);
-			});
-		});
-
-		describe('when the peer does not exist', () => {
-			it('should return false when the peer does not exist', async () => {
-				const success = triedPeersObj.updatePeer(samplePeers[2]);
-				expect(success).to.be.false;
-			});
-		});
-	});
-
 	describe('#failedConnectionAction', () => {
 		let triedPeersObj: TriedList;
 		const samplePeers = initializePeerInfoList();
